set blockchaininfo.pruned to false if fetch_blocks
