+++ Cargo.toml
@@ -23,21 +23,22 @@ spec = "config_spec.toml"
 compat = []
 
 [dependencies]
-anyhow = "1.0.33"
+anyhow = "1.0.34"
 base32 = "0.4.0"
 base64 = "0.13.0"
-bitcoin = { version = "0.25.0", features = ["use-serde"] }
+bitcoin = { version = "0.25.2", features = ["use-serde"] }
 configure_me = { version = "0.3.4" }
 derive_more = "0.99.11"
 enum_future = "0.1"
-futures = "0.3.6"
+futures = "0.3.8"
 hex = "0.4.2"
 http = "0.2.1"
-hyper = "0.13.8"
+hyper = "0.13.9"
 itertools = "0.9.0"
 lazy_static = "1.4.0"
+linear-map = { version = "1.2.0", features = ["serde_impl"] }
 rand = "0.7.3"
-serde = { version = "1.0.116", features = ["derive"] }
+serde = { version = "1.0.117", features = ["derive"] }
 serde_json = "1.0.59"
 slog = "2.5.2"
 slog-async = "2.5.0"
