+++ src/client.rs
@@ -1,14 +1,7 @@
-use std::collections::{HashMap, HashSet};
 use std::future::Future;
 use std::path::PathBuf;
 
 use anyhow::{anyhow, Error};
-use bitcoin::{
-    consensus::Decodable,
-    hash_types::BlockHash,
-    network::{constants::ServiceFlags, Address},
-    util::amount::Amount,
-};
 use futures::{channel::mpsc, StreamExt, TryStreamExt};
 use hyper::{
     body::Bytes,
@@ -22,8 +15,6 @@ use serde::{
     ser::{Serialize, Serializer},
 };
 use serde_json::Value;
-
-use crate::util::{Either, HexBytes};
 
 #[cfg(feature = "compat")]
 use crate::util::compat::StrCompat;
@@ -117,249 +108,6 @@ impl RpcMethod for GenericRpcMethod {
     type Response = Value;
     fn as_str<'a>(&'a self) -> &'a str {
         self.0.as_str()
-    }
-}
-
-#[derive(Debug)]
-pub struct GetBlock;
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub struct GetBlockParams(
-    pub BlockHash,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub Option<u8>,
-);
-impl RpcMethod for GetBlock {
-    type Params = GetBlockParams;
-    type Response = Either<HexBytes, GetBlockResult>;
-    fn as_str(&self) -> &'static str {
-        "getblock"
-    }
-}
-impl Serialize for GetBlock {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        self.as_str().serialize(serializer)
-    }
-}
-impl<'de> Deserialize<'de> for GetBlock {
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        let s: &'de str = Deserialize::deserialize(deserializer)?;
-        if s == Self.as_str() {
-            Ok(Self)
-        } else {
-            Err(serde::de::Error::invalid_value(
-                serde::de::Unexpected::Str(s),
-                &Self.as_str(),
-            ))
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct GetBlockHeader;
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub struct GetBlockHeaderParams(
-    pub BlockHash,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub Option<bool>,
-);
-impl RpcMethod for GetBlockHeader {
-    type Params = GetBlockHeaderParams;
-    type Response = Either<HexBytes, GetBlockHeaderResult>;
-    fn as_str(&self) -> &'static str {
-        "getblockheader"
-    }
-}
-impl Serialize for GetBlockHeader {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        self.as_str().serialize(serializer)
-    }
-}
-impl<'de> Deserialize<'de> for GetBlockHeader {
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        let s: &'de str = Deserialize::deserialize(deserializer)?;
-        if s == Self.as_str() {
-            Ok(Self)
-        } else {
-            Err(serde::de::Error::invalid_value(
-                serde::de::Unexpected::Str(s),
-                &Self.as_str(),
-            ))
-        }
-    }
-}
-
-#[derive(Debug, serde::Deserialize, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GetBlockHeaderResult {
-    pub hash: bitcoin::BlockHash,
-    pub confirmations: u32,
-    pub height: usize,
-    pub version: i32,
-    pub version_hex: Option<HexBytes>,
-    pub merkleroot: bitcoin::TxMerkleNode,
-    pub time: usize,
-    pub mediantime: Option<usize>,
-    pub nonce: u32,
-    pub bits: String,
-    pub difficulty: f64,
-    pub chainwork: HexBytes,
-    pub n_tx: usize,
-    pub previousblockhash: Option<bitcoin::BlockHash>,
-    pub nextblockhash: Option<bitcoin::BlockHash>,
-}
-
-#[derive(Debug, serde::Deserialize, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GetBlockResult {
-    #[serde(flatten)]
-    pub header: GetBlockHeaderResult,
-    pub size: usize,
-    pub strippedsize: Option<usize>,
-    pub weight: usize,
-    pub tx: Vec<bitcoin::Txid>,
-}
-
-#[derive(Debug)]
-pub struct GetPeerInfo;
-
-// https://docs.rs/bitcoincore-rpc-json/0.12.0/bitcoincore_rpc_json/struct.GetPeerInfoResult.html
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub struct PeerInfo {
-    /// Peer index
-    pub id: u64,
-    /// The IP address and port of the peer
-    pub addr: String,
-    /// Bind address of the connection to the peer
-    // TODO: use a type for addrbind
-    pub addrbind: String,
-    /// Local address as reported by the peer
-    // TODO: use a type for addrlocal
-    pub addrlocal: Option<String>,
-    /// The services offered
-    // TODO: use a type for services
-    pub services: String,
-    /// The services offered
-    pub servicesnames: HashSet<String>,
-    /// Whether peer has asked us to relay transactions to it
-    pub relaytxes: bool,
-    /// The time in seconds since epoch (Jan 1 1970 GMT) of the last send
-    pub lastsend: u64,
-    /// The time in seconds since epoch (Jan 1 1970 GMT) of the last receive
-    pub lastrecv: u64,
-    /// The total bytes sent
-    pub bytessent: u64,
-    /// The total bytes received
-    pub bytesrecv: u64,
-    /// The connection time in seconds since epoch (Jan 1 1970 GMT)
-    pub conntime: u64,
-    /// The time offset in seconds
-    pub timeoffset: i64,
-    /// ping time (if available)
-    pub pingtime: Option<f64>,
-    /// minimum observed ping time (if any at all)
-    pub minping: Option<f64>,
-    /// ping wait (if non-zero)
-    pub pingwait: Option<f64>,
-    /// The peer version, such as 70001
-    pub version: u64,
-    /// The string version
-    pub subver: String,
-    /// Inbound (true) or Outbound (false)
-    pub inbound: bool,
-    /// Whether connection was due to `addnode`/`-connect` or if it was an
-    /// automatic/inbound connection
-    pub addnode: bool,
-    /// The starting height (block) of the peer
-    pub startingheight: i64,
-    /// The ban score
-    pub banscore: i64,
-    /// The last header we have in common with this peer
-    pub synced_headers: i64,
-    /// The last block we have in common with this peer
-    pub synced_blocks: i64,
-    /// The heights of blocks we're currently asking from this peer
-    pub inflight: Vec<u64>,
-    /// Whether the peer is whitelisted
-    pub whitelisted: bool,
-    #[serde(
-        rename = "minfeefilter",
-        default,
-        with = "bitcoin::util::amount::serde::as_btc::opt"
-    )]
-    pub min_fee_filter: Option<Amount>,
-    /// The total bytes sent aggregated by message type
-    pub bytessent_per_msg: HashMap<String, u64>,
-    /// The total bytes received aggregated by message type
-    pub bytesrecv_per_msg: HashMap<String, u64>,
-}
-impl PeerInfo {
-    pub fn into_address(self) -> Result<Address, Error> {
-        let services = ServiceFlags::consensus_decode(&mut std::io::Cursor::new(hex::decode(
-            &self.services,
-        )?))?;
-        if let Ok(sock_addr) = self.addr.parse() {
-            Ok(Address::new(&sock_addr, services))
-        } else {
-            let mut addr_split = self.addr.split(":");
-            let host = addr_split
-                .next()
-                .ok_or_else(|| anyhow!("Invalid Peer Address: {}", self.addr))?;
-            let port = addr_split
-                .next()
-                .ok_or_else(|| anyhow!("Invalid Peer Address: {}", self.addr))?
-                .parse()?;
-            let onion = base32::decode(
-                base32::Alphabet::RFC4648 { padding: false },
-                host.strip_suffix(".onion")
-                    .ok_or_else(|| anyhow!("Invalid Peer Address: {}", self.addr))?,
-            )
-            .ok_or_else(|| anyhow!("Invalid Peer Address: {}", self.addr))?;
-            if onion.len() < 10 {
-                return Err(anyhow!("Invalid Peer Address: {}", self.addr));
-            }
-            let address: [u16; 8] = [
-                0xFD87,
-                0xD87E,
-                0xEB43,
-                ((onion[0] as u16) << 8) + (onion[1] as u16),
-                ((onion[2] as u16) << 8) + (onion[3] as u16),
-                ((onion[4] as u16) << 8) + (onion[5] as u16),
-                ((onion[6] as u16) << 8) + (onion[7] as u16),
-                ((onion[8] as u16) << 8) + (onion[9] as u16),
-            ];
-            Ok(Address {
-                services,
-                address,
-                port,
-            })
-        }
-    }
-}
-impl RpcMethod for GetPeerInfo {
-    type Params = [(); 0];
-    type Response = Vec<PeerInfo>;
-    fn as_str(&self) -> &'static str {
-        "getpeerinfo"
-    }
-}
-impl Serialize for GetPeerInfo {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        self.as_str().serialize(serializer)
-    }
-}
-impl<'de> Deserialize<'de> for GetPeerInfo {
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        let s: &'de str = Deserialize::deserialize(deserializer)?;
-        if s == Self.as_str() {
-            Ok(Self)
-        } else {
-            Err(serde::de::Error::invalid_value(
-                serde::de::Unexpected::Str(s),
-                &Self.as_str(),
-            ))
-        }
     }
 }
